+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 ### Added
 - `Reference.get_polygons`, `Reference.get_paths`, `Reference.get_labels`
+- `Library.rename_cell`
 - `Library::rename_cell` and `Library::replace_cell` in the C++ API
 ### Changed
 - `Cell.filter` arguments modified to match `read_gds`.